6-size.c

+++ 0x00-hello_world/6-size.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 /**
- * main - A program that prints the size of various computer types
- * Return: 0 (success)
+ * main - c program that prints the size of various computer types
+ * Return: 0
  */
 int main(void)
 {
@@ -10,10 +10,10 @@ int main(void)
 	long int c;
 	long long int d;
 	float f;
-printf("size of a char: %lu byte(s)\n", (unsigned long)sizeof(a));
-printf("size of an int: %lu byte(s)\n", (unsigned long)sizeof(b));
-printf("size of a long int: %lu byte(s)\n", (unsigned long)sizeof(c));
-printf("size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(d));
-printf("size of a float: %lu byte(s)\n", (unsigned long)sizeof(f));
+printf("size of a char: %ld byte(s)\n", sizeof(a));
+printf("size of an int: %ld byte(s)\n", sizeof(b));
+printf("size of a long int: %ld byte(s)\n", sizeof(c));
+printf("size of a long long int: %ld byte(s)\n", sizeof(d));
+printf("size of a float: %ld byte(s)\n", sizeof(f));
 return (0);
 }
